requirements.txt
PyQt5>=5.15.0
matplotlib>=3.3.0
numpy>=1.19.0
numpy-stl>=2.16.0
